@@ -1,6 +1,6 @@
 {
     "name": "cashtab",
-    "version": "1.0.21",
+    "version": "1.0.22",
     "private": true,
     "scripts": {
         "start": "node scripts/start.js",
@@ -45,6 +45,7 @@
         "setupFiles": [
             "react-app-polyfill/jsdom"
         ],
+        "testTimeout": 10000,
         "testMatch": [
             "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
             "<rootDir>/src/**/*.{spec,test}.{js,jsx,ts,tsx}"
@@ -100,8 +101,8 @@
         "etoken-list": "^1.0.4",
         "localforage": "^1.9.0",
         "qrcode.react": "^3.1.0",
-        "react": "^17.0.2",
-        "react-dom": "^17.0.2",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
         "react-easy-crop": "^3.5.3",
         "react-ga": "^3.3.0",
         "react-image": "^4.0.3",
@@ -114,8 +115,7 @@
         "@pmmmwh/react-refresh-webpack-plugin": "^0.5.3",
         "@svgr/webpack": "^6.3.1",
         "@testing-library/jest-dom": "^5.16.2",
-        "@testing-library/react": "^12.1.3",
-        "@testing-library/react-hooks": "^8.0.1",
+        "@testing-library/react": "^14.2.0",
         "@testing-library/user-event": "^14.5.2",
         "assert": "^2.0.0",
         "babel-jest": "^27.4.2",
