@@ -1,19 +1,20 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import './index.css';
 import App from './components/App';
 import { WalletProvider } from './utils/context';
 import { HashRouter as Router } from 'react-router-dom';
 import GA from './utils/GoogleAnalytics';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container);
+root.render(
     <WalletProvider>
         <Router>
             {GA.init() && <GA.RouteTracker />}
             <App />
         </Router>
     </WalletProvider>,
-    document.getElementById('root'),
 );
 
 if (module.hot) {
